+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- None
+- Removed dependencies to scheduled_updates module in degov_node_press
+- Added content_moderation and workflows core modules dependency to degov.info.yml
 
 ## [2.0.11] - 27-07-2018
 - Improved namespaces regarding Behat testing